requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
kafka-python==2.0.2
py4j==0.10.9.7
pyspark==3.5.1
requests==2.31.0
urllib3==2.2.1
